refactor: extract socket module with layout as submodule
Move `SocketKind`, `Socket`, `PositionedSocket`, `SocketResponses` (+ impl)
from `node.rs` into `src/socket/mod.rs`. Extract the Phase A (highlight state
extraction) and Phase B (layer setup, painting, `ui.interact()`) socket code
from `show_impl` into `socket::show`.

Relocate the former `socket_layout` module into `socket::layout`, grouping all
socket-related code under a single `socket` module.

Update `lib.rs` re-exports and `examples/demo.rs` import paths accordingly.
